% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_weights_matrix.R
\name{create_test_matrix}
\alias{create_test_matrix}
\title{Transforms a vector of text to the same format of a given training matrix (Retain same words)}
\usage{
create_test_matrix(Text, training_matrix, ngrams = 1, verbose = FALSE)
}
\arguments{
\item{Text}{Text to be adapted}

\item{training_matrix}{The matrix that was used to train the model}

\item{ngrams}{Vector that specifies type of ngrams to generate}
}
\value{
An adapted matrix
}
\description{
Transforms a vector of text to the same format of a given training matrix (Retain same words)
}
