% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_train_test_matrix.R
\name{create_training_matrix}
\alias{create_training_matrix}
\title{Create a matrix from a vector of text. This function does not do any text cleaning.}
\usage{
create_training_matrix(text, labels, minDocFreq = 0, maxDocFreq = 1,
  minWordLength = 0, maxWordLength = Inf, weighting = "count",
  ngrams = 1, features = NULL, verbose = FALSE)
}
\arguments{
\item{text}{A vector of strings to be made into a matrix}

\item{labels}{Labels for classification}

\item{minDocFreq}{Minimum frequency of a word to be included (count or fraction)}

\item{maxDocFreq}{Maxmimum frequency of a word to be included (count or fraction)}

\item{minWordLength}{Minimum wordlength for a word to be included}

\item{maxWordLength}{Maximum wordlength for a word to be included}

\item{weighting}{Weighting to use. Choose between frequency, relFreq, relMaxFreq, logFreq, tfidf (See dfm_weight)}

\item{ngrams}{Vector that specifies type of ngrams to generate}
}
\value{
A document feature matrix
}
\description{
Create a matrix from a vector of text. This function does not do any text cleaning.
}
