% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_weights_matrix.R
\name{create_weights_matrix}
\alias{create_weights_matrix}
\title{Create a weights matrix given a MaxEnt model
Given a matrix with labels and probabilities}
\usage{
create_weights_matrix(matrix, model)
}
\arguments{
\item{matrix}{A matrix}

\item{model}{A MaxEnt model}
}
\value{
A weight matrix data table
}
\description{
Create a weights matrix given a MaxEnt model
Given a matrix with labels and probabilities
}
