% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choose_relevant_words.R
\name{get_common_words}
\alias{get_common_words}
\title{Returns common words for a matrix}
\usage{
get_common_words(text, labels, ngrams = 1, minDocFreq = 0.1, batches = 1)
}
\arguments{
\item{text}{A vector of strings to be made into a matrix}
}
\value{
A document feature matrix
}
\description{
Returns common words for a matrix
}
