% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choose_relevant_words.R
\name{get_common_words_single}
\alias{get_common_words_single}
\title{Returns common words for a matrix}
\usage{
get_common_words_single(text, minDocFreq = 1, maxDocFreq = Inf,
  minWordLength = 0, maxWordLength = Inf, weighting = "frequency",
  ngrams = 1, minCount = 1, verbose = FALSE)
}
\arguments{
\item{text}{A vector of strings to be made into a matrix}

\item{minDocFreq}{Minimum frequency of a word to be included (count or fraction)}

\item{maxDocFreq}{Maxmimum frequency of a word to be included (count or fraction)}

\item{minWordLength}{Minimum wordlength for a word to be included}

\item{maxWordLength}{Maximum wordlength for a word to be included}

\item{weighting}{Weighting to use. Choose between frequency, relFreq, relMaxFreq, logFreq, tfidf (See dfm_weight)}

\item{ngrams}{Vector that specifies type of ngrams to generate}
}
\value{
A document feature matrix
}
\description{
Returns common words for a matrix
}
