% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MaxEntClassifier.R
\name{predictMC}
\alias{predictMC}
\title{Predict using a previously trained classifier, given a matrix with labels and probabilities}
\usage{
predictMC(model, feature_matrix, probabilities = -1, batches = 1)
}
\arguments{
\item{model}{A MaxEnt model}

\item{feature_matrix}{The basis for prediction}

\item{probabilities}{{-1, 0, 1} {Return probabilities for all classes / only labels / only the top class}}

\item{batches}{In how many batches to split the dataset. This uses multithreading.}
}
\value{
A data frame with predictions
}
\description{
Predict using a previously trained classifier, given a matrix with labels and probabilities
}
