% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MaxEntClassifier.R
\name{score}
\alias{score}
\title{Scores the performance of a model}
\usage{
score(predicted, real)
}
\arguments{
\item{predicted}{Labels that were predicted}

\item{real}{Actual labels}
}
\value{
Accuracy
}
\description{
Scores the performance of a model
}
