% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MaxEntClassifier.R
\name{trainMC}
\alias{trainMC}
\title{Train a classifier
Given a matrix with word counts and a vector of labels, train a MaxEnt classifier}
\usage{
trainMC(feature_matrix, code_vector, l1_regularizer = 0, l2_regularizer = 0,
  use_sgd = FALSE, set_heldout = 0, verbose = FALSE)
}
\arguments{
\item{train_matrix}{A matrix containing word counts}

\item{train_labels}{A matrix containing the labels for the training matrix}
}
\value{
A maxent model
}
\description{
Train a classifier
Given a matrix with word counts and a vector of labels, train a MaxEnt classifier
}
