% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tune_maxent.R
\name{tune_MaxEnt}
\alias{tune_MaxEnt}
\title{Tunes a MaxEnt classifier)}
\usage{
tune_MaxEnt(feature_matrix, feature_labels, nfold = 3, showall = TRUE,
  verbose = TRUE, cores = 1)
}
\arguments{
\item{feature_matrix}{A matrix with features}

\item{feature_labels}{The labels for the matrix}

\item{nfold}{Number of folds for k-fold crossvalidation (k = 5 for basic run, k = 10 is accurate)}

\item{showall}{Show all of the configurations}

\item{verbose}{Output extra information}

\item{cores}{How many cores to use}
}
\value{
A matrix with the best parameters found
}
\description{
Tunes a MaxEnt classifier)
}
